Splitup server setup
